backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymssql==2.2.11
PyMySQL==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
cryptography==41.0.8